Add continuous locomotion (walking)
The VRPlayer is now a Spatial node. There is a KinematicBody node
as its child, with the ARVROrigin as a child of that. When walking,
the KinematicBody is moved with move_and_slide. The other nodes that
were part of the VRPlayer scene before are parented to the root
Spatial node, which is what is moved when teleporting or turning.

Teleporting and walking can both be enabled, but the default bindings
have to be changed so that teleporting and walking forward are on
different buttons.